requirements.txt
# Workout App Python Dependencies
# Install with: pip install -r requirements.txt

# Google API libraries
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.0

# Web framework
flask>=2.3.0
requests>=2.25.0          # For HTTP requests

# CLI utilities
tabulate>=0.9.0

# Data handling
pandas>=1.5.0        # Optional: for advanced data analysis

# Testing (optional)
pytest>=7.0.0        # For running tests
pytest-cov>=4.0.0    # For coverage reports
flask-cors>=4.0.0     # For frontend-backend communication
